Remove deprecated session search module
